move chunkers and hashers from lib to bin
