+++ src/chunker.rs
@@ -1,5 +1,3 @@
-use std::cmp::min;
-
 /// A chunk of the processed data. Doesn't store any data,
 /// only contains offset and length of the chunk.
 #[derive(Copy, Clone, Debug)]
@@ -9,13 +7,21 @@ pub struct Chunk {
 }
 
 impl Chunk {
-    pub(crate) fn new(offset: usize, length: usize) -> Self {
+    pub fn new(offset: usize, length: usize) -> Self {
         Self { offset, length }
     }
 
     /// Effective range of the chunk in the data.
-    pub(crate) fn range(&self) -> std::ops::Range<usize> {
+    pub fn range(&self) -> std::ops::Range<usize> {
         self.offset..self.offset + self.length
+    }
+
+    pub fn length(&self) -> usize {
+        self.length
+    }
+
+    pub fn offset(&self) -> usize {
+        self.offset
     }
 }
 
@@ -43,75 +49,4 @@ pub trait Chunker {
     /// data buffer. Used to pre-allocate the buffer with the required size so that allocation times are not counted
     /// towards total chunking time.
     fn estimate_chunk_count(&self, data: &[u8]) -> usize;
-}
-
-/// Chunker that utilizes Fixed Sized Chunking (FSC) algorithm,
-/// splitting file into even-sized chunks.
-#[derive(Debug)]
-pub struct FSChunker {
-    chunk_size: usize,
-    rest: Vec<u8>,
-}
-
-impl FSChunker {
-    pub fn new(chunk_size: usize) -> Self {
-        Self {
-            chunk_size,
-            rest: vec![],
-        }
-    }
-}
-
-impl Chunker for FSChunker {
-    fn chunk_data(&mut self, data: &[u8], empty: Vec<Chunk>) -> Vec<Chunk> {
-        let mut offset = 0;
-        let mut chunks = empty;
-        while offset < data.len() {
-            let chunk = Chunk::new(offset, min(self.chunk_size, data.len() - offset));
-            chunks.push(chunk);
-            offset += self.chunk_size;
-        }
-
-        let last_chunk = chunks.last().unwrap();
-        if last_chunk.length < self.chunk_size {
-            self.rest = data[last_chunk.range()].to_vec();
-        } else {
-            self.rest = vec![];
-        }
-        chunks
-    }
-
-    fn remainder(&self) -> &[u8] {
-        &self.rest
-    }
-
-    fn estimate_chunk_count(&self, data: &[u8]) -> usize {
-        data.len() / self.chunk_size + 1
-    }
-}
-
-#[derive(Default, Debug)]
-pub struct LeapChunker {
-    rest: Vec<u8>,
-}
-
-impl Chunker for LeapChunker {
-    fn chunk_data(&mut self, data: &[u8], empty: Vec<Chunk>) -> Vec<Chunk> {
-        let chunker = chunking::leap_based::Chunker::new(data);
-        let mut chunks = empty;
-        for chunk in chunker {
-            chunks.push(Chunk::new(chunk.pos, chunk.len));
-        }
-
-        self.rest = data[chunks.pop().unwrap().range()].to_vec();
-        chunks
-    }
-
-    fn remainder(&self) -> &[u8] {
-        &self.rest
-    }
-
-    fn estimate_chunk_count(&self, data: &[u8]) -> usize {
-        data.len() / 1024 * 8
-    }
 }
